src/main/java/com/stackroute/MovieApp/repository/MovieRepository.java
package com.stackroute.MovieApp.repository;

import com.stackroute.MovieApp.domain.Movie;
import com.stackroute.MovieApp.exceptions.MovieNotFoundException;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
public interface MovieRepository extends JpaRepository<Movie, Integer> {

    @Query("SELECT m FROM Movie  m WHERE m.movieName = :searchName")
    public List<Movie> findByName(@Param("searchName") String searchName) throws MovieNotFoundException;

}
